Draw-348 refactor: SessionUseCase 이관
