--- app/websocket/src/main/kotlin/com/xorker/draw/websocket/EmptyObject.kt
@@ -1,3 +1,0 @@
-package com.xorker.draw.websocket
-
-object EmptyObject
