+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/log/WebSocketLogger.kt
@@ -6,9 +6,9 @@ import com.xorker.draw.support.logging.logger
 import com.xorker.draw.support.logging.registerRequestId
 import com.xorker.draw.websocket.SessionId
 import com.xorker.draw.websocket.SessionInitializeRequest
-import com.xorker.draw.websocket.SessionUseCase
 import com.xorker.draw.websocket.message.request.RequestAction
 import com.xorker.draw.websocket.message.request.dto.WebSocketRequest
+import com.xorker.draw.websocket.session.SessionManager
 import org.slf4j.MDC
 import org.springframework.stereotype.Component
 import org.springframework.web.socket.CloseStatus
@@ -16,7 +16,7 @@ import org.springframework.web.socket.WebSocketSession
 
 @Component
 internal class WebSocketLogger(
-    private val sessionUseCase: SessionUseCase,
+    private val sessionManager: SessionManager,
     private val objectMapper: ObjectMapper,
 ) {
     private val logger = logger()
@@ -80,7 +80,7 @@ internal class WebSocketLogger(
     }
 
     private fun injectDefaultSessionInfo(sessionId: SessionId, data: MutableMap<String, Any?>) {
-        val sessionDto = sessionUseCase.getSession(sessionId)
+        val sessionDto = sessionManager.getSession(sessionId)
         if (sessionDto != null) {
             data["userId"] = sessionDto.user.id.value
             data["roomId"] = sessionDto.roomId
