+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/session/SessionFactory.kt
@@ -1,4 +1,4 @@
-package com.xorker.draw.websocket
+package com.xorker.draw.websocket.session
 
 import com.xorker.draw.auth.token.TokenUseCase
 import com.xorker.draw.exception.UnAuthenticationException
@@ -6,6 +6,10 @@ import com.xorker.draw.room.RoomId
 import com.xorker.draw.room.RoomRepository
 import com.xorker.draw.user.User
 import com.xorker.draw.user.UserId
+import com.xorker.draw.websocket.MafiaGameRandomMatchingRequest
+import com.xorker.draw.websocket.Session
+import com.xorker.draw.websocket.SessionInitializeRequest
+import com.xorker.draw.websocket.WaitingQueueSession
 import com.xorker.draw.websocket.message.request.SessionWrapper
 import com.xorker.draw.websocket.message.request.WaitingQueueSessionWrapper
 import org.slf4j.MDC
