+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/session/SessionManager.kt
@@ -1,6 +1,10 @@
-package com.xorker.draw.websocket
+package com.xorker.draw.websocket.session
 
 import com.xorker.draw.user.UserId
+import com.xorker.draw.websocket.Session
+import com.xorker.draw.websocket.SessionEventListener
+import com.xorker.draw.websocket.SessionId
+import com.xorker.draw.websocket.SessionInitializeRequest
 import java.util.concurrent.ConcurrentHashMap
 import org.springframework.core.Ordered
 import org.springframework.core.annotation.Order
@@ -8,30 +12,15 @@ import org.springframework.stereotype.Service
 
 @Order(Ordered.HIGHEST_PRECEDENCE)
 @Service
-internal class SessionService : SessionUseCase, SessionEventListener {
+internal class SessionManager : SessionEventListener {
     private val sessionMap: ConcurrentHashMap<SessionId, Session> = ConcurrentHashMap()
     private val userIdMap: ConcurrentHashMap<UserId, Session> = ConcurrentHashMap()
 
-    override fun registerSession(session: Session) {
-        if (sessionMap.contains(session.id)) {
-            // Init을 중복으로 호출 하면 기존 데이터를 Unregister 하고 Init 한다.
-            unregisterSession(session.id)
-        }
-
-        sessionMap[session.id] = session
-        userIdMap[session.user.id] = session
-    }
-
-    override fun unregisterSession(sessionId: SessionId) {
-        val session = sessionMap.remove(sessionId)
-        userIdMap.remove(session?.user?.id)
-    }
-
-    override fun getSession(sessionId: SessionId): Session? {
+    fun getSession(sessionId: SessionId): Session? {
         return sessionMap[sessionId]
     }
 
-    override fun getSession(userId: UserId): Session? {
+    fun getSession(userId: UserId): Session? {
         return userIdMap[userId]
     }
 
@@ -49,5 +38,20 @@ internal class SessionService : SessionUseCase, SessionEventListener {
 
     override fun exitSession(session: Session) {
         unregisterSession(session.id)
+    }
+
+    private fun registerSession(session: Session) {
+        if (sessionMap.contains(session.id)) {
+            // Init을 중복으로 호출 하면 기존 데이터를 Unregister 하고 Init 한다.
+            unregisterSession(session.id)
+        }
+
+        sessionMap[session.id] = session
+        userIdMap[session.user.id] = session
+    }
+
+    private fun unregisterSession(sessionId: SessionId) {
+        val session = sessionMap.remove(sessionId)
+        userIdMap.remove(session?.user?.id)
     }
 }
